migrate to appDir
