+++ package.json
@@ -14,7 +14,7 @@
     "@tanstack/react-query": "^4.29.5",
     "@vercel/analytics": "^1.0.1",
     "disable-scroll": "^0.6.0",
-    "next": "^13.3.2",
+    "next": "^13.4.1",
     "nodemailer": "^6.9.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -24,17 +24,17 @@
   },
   "devDependencies": {
     "@types/eslint": "^8.37.0",
-    "@types/node": "^18.16.3",
+    "@types/node": "^20.1.0",
     "@types/nodemailer": "^6.4.7",
     "@types/prettier": "^2.7.2",
-    "@types/react": "18.2.1",
-    "@types/react-dom": "^18.2.2",
+    "@types/react": "18.2.5",
+    "@types/react-dom": "^18.2.4",
     "@types/react-google-recaptcha": "^2.1.5",
     "@typescript-eslint/eslint-plugin": "^5.59.2",
     "@typescript-eslint/parser": "^5.59.2",
     "autoprefixer": "^10.4.14",
-    "eslint": "^8.39.0",
-    "eslint-config-next": "^13.3.2",
+    "eslint": "^8.40.0",
+    "eslint-config-next": "^13.4.1",
     "postcss": "^8.4.23",
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.2.8",
