--- public/robots.txt
@@ -1,9 +1,0 @@
-# *
-User-agent: *
-Allow: /
-
-# Host
-Host: https://znagy.hu
-
-# Sitemaps
-Sitemap: https://znagy.hu/sitemap.xml
