--- public/sitemap.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<urlset
-      xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9
-            http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd">
-<url>
-  <loc>https://znagy.hu/</loc>
-  <priority>1.00</priority>
-</url>
-<url>
-  <loc>https://znagy.hu/contact</loc>
-  <priority>0.80</priority>
-</url>
-<url>
-  <loc>https://znagy.hu/setup</loc>
-  <priority>0.80</priority>
-</url>
-<url>
-  <loc>https://znagy.hu/resume</loc>
-  <priority>0.80</priority>
-</url>
-</urlset>
