--- src/components/Home/index.ts
@@ -1,6 +1,0 @@
-import { AboutMe } from './AboutMe';
-import { GetInTouch } from './GetInTouch';
-import { Hero } from './Hero';
-import { Projects } from './Projects';
-
-export { AboutMe, GetInTouch, Projects, Hero };
