+++ src/components/ReactQueryWrapper.tsx
@@ -1,0 +1,16 @@
+'use client';
+
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      staleTime: 60,
+      refetchOnWindowFocus: false,
+    },
+  },
+});
+
+export const ReactQueryWrapper: React.FC<{ children: React.ReactNode }> = ({ children }) => (
+  <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+);
