--- src/pages/_document.tsx
@@ -1,15 +1,0 @@
-import type { DocumentType } from 'next/dist/shared/lib/utils';
-import { Html, Head, Main, NextScript } from 'next/document';
-
-const Document: DocumentType = () => (
-  <Html lang="en">
-    <Head />
-
-    <body>
-      <Main />
-      <NextScript />
-    </body>
-  </Html>
-);
-
-export default Document;
